Refactoring commands into seperate file
