+++ shell/commands_test.go
